Added a new color class that I will use everywhere.  It includes a large predefined set of colors.  This is now integrated with the new console work.
Started adding a congle Console class that will merge the different files from the previous implementation so we can do more advanced behaviors without having to constantly modify and remember so many classes in other parts of the code.

+++ RPGEngine/Source/Terminal/sdl_context.cpp
@@ -10,7 +10,7 @@ namespace Term
 {
 	namespace SDL
 	{
-		Context::Context(const size_t width, const size_t height) :
+		Context::Context(const int width, const int height) :
 			twidth(0), theight(0),
 			tilemap(nullptr),
 			tilemap_surface(nullptr),
